refactor(User): 사용자 클래스 분리
